Fix 8 crates with API drift compilation errors
- aspen-testing-core: ScanRequest.limit → limit_results, ScanResult.count → result_count
- aspen-sharding: ScanRequest.limit → limit_results
- aspen-raft: ScanRequest.limit → limit_results in node tests
- aspen-federation: ResourceMetadata.threshold → threshold_replicas in sync tests
- aspen-secrets: Rewrite InMemoryKvStore mock to match current KeyValueStore trait
- aspen-secrets-handler: SecretsError::ValueTooLarge field names (size→size_bytes, max→max_bytes)
- aspen-blob: OsRng API change (rand::rngs::OsRng → rand::rng())
- aspen-docs: LOG_BROADCAST_BUFFER_SIZE u32→usize cast for broadcast::channel

+++ crates/aspen-docs/src/exporter/mod.rs
@@ -94,7 +94,7 @@ mod tests {
         let exporter = Arc::new(DocsExporter::new(writer.clone()));
 
         // Create broadcast channel (same as in bootstrap)
-        let (sender, _) = broadcast::channel::<LogEntryPayload>(LOG_BROADCAST_BUFFER_SIZE);
+        let (sender, _) = broadcast::channel::<LogEntryPayload>(LOG_BROADCAST_BUFFER_SIZE as usize);
 
         // Subscribe and spawn exporter
         let receiver = sender.subscribe();
@@ -136,7 +136,7 @@ mod tests {
         let writer = Arc::new(InMemoryDocsWriter::new());
         let exporter = Arc::new(DocsExporter::new(writer.clone()));
 
-        let (sender, _) = broadcast::channel::<LogEntryPayload>(LOG_BROADCAST_BUFFER_SIZE);
+        let (sender, _) = broadcast::channel::<LogEntryPayload>(LOG_BROADCAST_BUFFER_SIZE as usize);
         let receiver = sender.subscribe();
         let cancel = exporter.clone().spawn(receiver);
 
@@ -189,7 +189,7 @@ mod tests {
         let writer = Arc::new(InMemoryDocsWriter::new());
         let exporter = Arc::new(DocsExporter::new(writer.clone()));
 
-        let (sender, _) = broadcast::channel::<LogEntryPayload>(LOG_BROADCAST_BUFFER_SIZE);
+        let (sender, _) = broadcast::channel::<LogEntryPayload>(LOG_BROADCAST_BUFFER_SIZE as usize);
         let receiver = sender.subscribe();
         let cancel = exporter.clone().spawn(receiver);
